Some services improvement
