+++ src/ServiceManager.ts
@@ -1,3 +1,10 @@
+/**
+ * Manages services
+ *
+ * Services help to decouple code. Use them sparingly to provide low-level functionality to other parts of the game.
+ * They should be added once at the start of the game and then not changed. Generally try to avoid passing the service
+ * manager around, instead fetch the services you need and then inject them into the classes that need them.
+ */
 export interface ServiceManager {
   /**
    * Add a service
@@ -15,39 +22,3 @@ export interface ServiceManager {
    */
   get<T extends {}>(s: string | symbol): T;
 }
-
-class ServiceManagerInternal implements ServiceManager {
-  private services = new Map<string | symbol, {}>();
-
-  /**
-   * Add a service
-   *
-   * @param s the symbol used to lookup the service
-   * @param service the service to add
-   */
-  public add(s: string | symbol, service: {}): void {
-    this.services.set(s, service);
-  }
-
-  /**
-   * Get a service
-   *
-   * @param s the symbol used to lookup the service
-   * @returns the service
-   */
-  public get<T extends {}>(s: string | symbol): T {
-    const service = this.services.get(s);
-    if (!service) {
-      throw new Error(`Service ${s.toString()} not found`);
-    }
-
-    return service as T;
-  }
-}
-
-export const GameServiceManager = ServiceManagerInternal;
-
-/**
- * ServiceManager is a singleton that holds all services
- */
-export const ServiceManager = new ServiceManagerInternal();
